Moved store, mapper and model classes to its own package
